requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   jsonschema
    #   outcome
    #   referencing
    #   trio
blinker==1.9.0
    # via streamlit
cachetools==6.2.4
    # via streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
    #   selenium
cffi==2.0.0
    # via trio
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via openai
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.14.0
    # via altair
numpy==2.4.0
    # via
    #   pandas
    #   pydeck
    #   streamlit
openai==2.14.0
    # via -r requirements.in
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via
    #   altair
    #   streamlit
    #   webdriver-manager
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.0.0
    # via streamlit
protobuf==6.33.2
    # via streamlit
pyarrow==22.0.0
    # via streamlit
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   webdriver-manager
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   streamlit
    #   webdriver-manager
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
selenium==4.39.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   openai
    #   trio
sortedcontainers==2.4.0
    # via trio
streamlit==1.52.2
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.1
    # via openai
trio==0.32.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.15.0
    # via
    #   altair
    #   openai
    #   pydantic
    #   pydantic-core
    #   selenium
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3[socks]==2.6.2
    # via
    #   requests
    #   selenium
watchdog==6.0.0
    # via streamlit
webdriver-manager==4.0.2
    # via -r requirements.in
websocket-client==1.9.0
    # via selenium
wsproto==1.3.2
    # via trio-websocket
